Fix Render LangGraph startup
