+++ app/agent/lc_tools.py
@@ -2,15 +2,15 @@
 from __future__ import annotations
 import json
 import os
-from typing import Any
 from pydantic import BaseModel, Field
-from langchain.tools import tool
-from tavily import TavilyClient
+from langchain_core.tools import tool
 
 try:
+    from app.services.web_search import search_web
     from app.services.web_fetch import fetch_and_clean
     from app.services.vectorstore import load_faiss_or_none
 except ModuleNotFoundError:
+    from services.web_search import search_web
     from services.web_fetch import fetch_and_clean
     from services.vectorstore import load_faiss_or_none
 
@@ -61,52 +61,6 @@ def retrieve_local_tool(query: str, k: int = 6) -> str:
 # -----------------------------
 # Web Search (Tavily) Tool
 # -----------------------------
-# This tool emits JSON-like search results (title, url, content) by default.
-# Requires TAVILY_API_KEY in the environment.
-def search_web(query: str, max_results: int = 3) -> dict[str, Any]:
-    """Return a stable, structured Tavily result contract."""
-    api_key = os.getenv("TAVILY_API_KEY")
-    if not api_key:
-        return {
-            "query": query,
-            "results": [],
-            "answer": None,
-            "error": "missing_tavily_api_key",
-        }
-    try:
-        client = TavilyClient(api_key=api_key)
-        resp = client.search(
-            query=query,
-            max_results=max(1, min(max_results, 5)),
-            include_answer=True,
-            include_raw_content=False,
-        )
-        results = []
-        for item in (resp or {}).get("results") or []:
-            results.append(
-                {
-                    "title": str(item.get("title") or "Web result"),
-                    "url": str(item.get("url") or ""),
-                    "content": str(item.get("content") or ""),
-                    "score": item.get("score"),
-                    "published_date": item.get("published_date"),
-                }
-            )
-        return {
-            "query": query,
-            "results": results,
-            "answer": (resp or {}).get("answer"),
-            "error": None,
-        }
-    except Exception as exc:
-        return {
-            "query": query,
-            "results": [],
-            "answer": None,
-            "error": f"{type(exc).__name__}: {exc}",
-        }
-
-
 @tool("tavily_search")
 def tavily_search_tool(query: str) -> str:
     """Search the web via Tavily and return structured JSON."""
