+++ pyproject.toml
@@ -8,12 +8,13 @@ dependencies = [
     "beautifulsoup4>=4.12",
     "faiss-cpu>=1.8.0",
     "fastapi>=0.115",
-    "langchain>=1.2.0,<2.0.0",
+    "langchain>=1.2.10,<1.3.0",
     "langchain-community>=0.4.0,<1.0.0",
     "langchain-openai>=1.1.0,<2.0.0",
     "langchain-tavily>=0.2.0,<1.0.0",
     "langchain-text-splitters>=1.1.0,<2.0.0",
-    "langgraph>=1.0.0,<2.0.0",
+    "langgraph==1.0.10",
+    "langgraph-prebuilt==1.0.8",
     "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
     "langgraph-cli>=0.4.11",
     "langgraph-sdk>=0.3.0,<1.0.0",
