+++ requirements.txt
@@ -12,7 +12,8 @@ langchain>=1.2.10,<1.3.0
 langchain-community>=0.3.0,<1.0.0
 langchain-openai>=1.1.9,<2.0.0
 langchain-text-splitters>=1.1.0,<2.0.0
-langgraph>=1.0.10,<1.1.0
+langgraph==1.0.10
+langgraph-prebuilt==1.0.8
 langgraph-sdk>=0.3.5,<1.0.0
 langgraph-checkpoint-sqlite>=1.0.0,<4.0.0
 langsmith>=0.1.94
